sync product supply

+++ app/models/supply.rb
@@ -40,28 +40,29 @@ class Supply < ApplicationRecord
   def copy_product_attr
     product = self.product
     shop = self.shop
-    self.name = product&.name
-    self.cost = product.cus_cost
-    self.cost_epub = product.cus_epub
-    
-    if shop.global_setting_enable == true
-      self.price = (self.cost * shop.cost_rate + self.cost_epub * shop.shipping_rate).round(2)
-    else
-      self.price = product&.suggest_price
-    end
+    if shop.present?
+      self.name = product&.name
+      self.cost = product.cus_cost
+      self.cost_epub = product.cus_epub
+      
+      if shop.global_setting_enable == true
+        self.price = (self.cost * shop.cost_rate + self.cost_epub * shop.shipping_rate).round(2)
+      else
+        self.price = product&.suggest_price
+      end
 
-    self.desc = product&.desc
-    random = rand(shop.random_from .. shop.random_to)
-    self.compare_at_price = (self.price * random/ 5).round(0) * 5
-    self.epub = (1 - shop.shipping_rate)*product.cus_epub
-    
-    self.supply_variants.destroy_all
-    product.variants.each do |variant|
+      self.desc = product&.desc
       random = rand(shop.random_from .. shop.random_to)
-      compare_at_price = (variant.price * random/ 5).round(0) * 5
-      self.supply_variants.create(option1: variant.option1, option2: variant.option2, option3: variant.option3, price: variant.price, sku: variant.sku, compare_at_price: compare_at_price)
+      self.compare_at_price = (self.price * random/ 5).round(0) * 5
+      self.epub = (1 - shop.shipping_rate)*product.cus_epub
+      
+      self.supply_variants.destroy_all
+      product.variants.each do |variant|
+        random = rand(shop.random_from .. shop.random_to)
+        compare_at_price = (variant.price * random/ 5).round(0) * 5
+        self.supply_variants.create(option1: variant.option1, option2: variant.option2, option3: variant.option3, price: variant.price, sku: variant.sku, compare_at_price: compare_at_price)
+      end
     end
-    #only create or update at product
 
     # TODO sync images
     # self.images = product.images
